clean up types and create gpp-utils
